refactor: rename base classes and client files for consistency
