+++ src/jira2py/api/issue_fields/issue_fields_async.py
@@ -6,7 +6,7 @@ from pydantic import validate_call
 
 from jira2py.client import JiraClientAsync, JiraCredentials
 
-from .issue_fields import IssueFieldsBase
+from .issue_fields_base import IssueFieldsBase
 
 
 class IssueFieldsAsync(IssueFieldsBase):
