+++ src/jira2py/api/jira_api_sync.py
@@ -4,7 +4,7 @@ from .issue_comments import IssueComments
 from .issue_fields import IssueFields
 from .issue_search import IssueSearch
 from .issues import Issues
-from .jira_api import JiraAPIBase
+from .jira_api_base import JiraAPIBase
 
 
 class JiraAPI(JiraAPIBase):
